requirements.txt
django
psycopg2-binary
djangorestframework
python-decouple
drf_yasg
djangorestframework-simplejwt